har fuckings lagt inn level check på alt ass

+++ BPO/app/Http/Controllers/RoomController.php
@@ -27,43 +27,56 @@ class RoomController extends Controller
     //lagre rom
     public function store(Request $request)
     {
-        $this->validate($request, [
-            'Rom' => 'required|alpha_num|max:45'
-        ]);
-        
-        //legg til rom
-        $rom_finnes = DB::select('SELECT * FROM room WHERE room = :request', ['request' => $request->Rom]);
-        if(!$rom_finnes)
+        if(session('levell') >= 2)
         {
-            DB::insert('INSERT INTO ROOM (room) VALUES (:request)',['request' => $request->Rom]);
-
-            \LogHelper::Log("Rom: ".$request->Rom, "1");
-
-            return redirect('/room')->with('success', 'Rom registrert');
-        }   
+            $this->validate($request, [
+                'Rom' => 'required|alpha_num|max:45'
+            ]);
+            
+            //legg til rom
+            $rom_finnes = DB::select('SELECT * FROM room WHERE room = :request', ['request' => $request->Rom]);
+            if(!$rom_finnes)
+            {
+                DB::insert('INSERT INTO ROOM (room) VALUES (:request)',['request' => $request->Rom]);
+    
+                \LogHelper::Log("Rom: ".$request->Rom, "1");
+    
+                return redirect('/room')->with('success', 'Rom registrert');
+            }   
+            else
+                return redirect('/room')->with('error', 'Rom finnes alt');
+        }
         else
-            return redirect('/room')->with('error', 'Rom finnes alt');
+        {
+            return redirect('/login')->with('error', 'Du er ikke admin og har ikke tilgang');
+        }
     }
 
     //slett rom
     public function destroy($id)
     {
-        $ibruk = DB::select('SELECT presentation_room FROM presentation, room 
-        WHERE presentation.presentation_room = room.room 
-        AND presentation.presentation_room = :id', ['id' => $id]);
-
-        if ($ibruk)
+        if(session('levell') >= 2)
         {
-            return redirect('/room')->with('error', 'Rommet er allerede i bruk og må fjernes fra presentasjonsplanen, før det kan slettes');
-        }
-        else{
-            DB::delete('DELETE FROM room WHERE room = :id', ['id' => $id]);
+            $ibruk = DB::select('SELECT presentation_room FROM presentation, room 
+            WHERE presentation.presentation_room = room.room 
+            AND presentation.presentation_room = :id', ['id' => $id]);
 
-            $bruker = session('navn');
-            \LogHelper::Log($bruker." har slettet rom med id ".$id, "1");
+            if ($ibruk)
+            {
+                return redirect('/room')->with('error', 'Rommet er allerede i bruk og må fjernes fra presentasjonsplanen, før det kan slettes');
+            }
+            else{
+                DB::delete('DELETE FROM room WHERE room = :id', ['id' => $id]);
 
-            return redirect('/room')->with('success', 'Rom Fjernet');
+                $bruker = session('navn');
+                \LogHelper::Log($bruker." har slettet rom med id ".$id, "1");
+
+                return redirect('/room')->with('success', 'Rom Fjernet');
+            }
         }
-        
+        else
+        {
+            return redirect('/login')->with('error', 'Du er ikke admin og har ikke tilgang');
+        }
     }
 }
